changed	amount to float

+++ vm-manager-roo/src/main/java/com/sjsu/minishare/model/UserCredit.java
@@ -26,7 +26,7 @@ public class UserCredit {
 
     private Integer totalCreditsUsed;
     
-    private Integer amount;
+    private float amount;
     
     private String paymentTransaction;
